+++ src/resources/views/auth/emails/password.blade.php
@@ -1,1 +1,1 @@
-{{ trans('backpack::base.click_here_to_reset') }}: <a href="{{ $link = url(config('backpack.base.route_prefix', 'admin').'/password/reset', $token).'?email='.urlencode($user->getEmailForPasswordReset()) }}"> {{ $link }} </a>
+{{ trans('backpack::base.click_here_to_reset') }}: <a href="{{ $link = backpack_url('password/reset', $token).'?email='.urlencode($user->getEmailForPasswordReset()) }}"> {{ $link }} </a>