+++ tests/BaseTest.php
@@ -1,7 +1,8 @@
 <?php
 
+use PHPUnit\Framework\TestCase as BaseTestCase;
 
-class BaseTest extends PHPUnit_Framework_TestCase
+class BaseTest extends BaseTestCase
 {
     /**
      * A basic test example.